+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Added a 3D speech bubble sample, for dynamically positioning a speech bubble above a game character. (@radiatoryang)
 - Added a phone chat sample. (@radiatoryang)
 - Added a visual novel template. (@radiatoryang)
 - Added support for voice overs. (@Schroedingers-Cat)